openpaser_server/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
torch>=2.0.0
transformers>=4.34.0
Pillow>=10.0.0
numpy>=1.24.0
httpx>=0.25.0 